fix delegation
Need to reslice positional args when delegating to another Delegator.
Otherwise, you can't access the help menus on Delegator Directives that
are direct children of the Root command. There's a slight deviation from
this on the Command.Run field. But the Command type only performs tasks,
it doesn't need to delegate (reslice args).