feat: Allow a Delegator to nest another Delegator
Removes the restriction on how deeply one can compose their CLI program.
Previously, a Delegator could not have other Delegators as a subcommand.
Attempting to select one of those resulted in an error:
        one does not simply create too many layers of delegators

These changes are an attempt to allow the developer to create an
arbitrary number of levels of delegators. Added some tests for one extra
level than before. This will probably cover most use cases. Haven't
tried creating more levels because it's not that immediately practical,
but will cross that bridge if it arrives.